Factor out Config object
